requirements-railway.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
aiosqlite
pandas
reportlab
Pillow
qrcode
openpyxl
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0
requests>=2.31.0
PyJWT>=2.8.0
